DEF[#79403]: 物联设备新增修改，modbus协议支持 modbus-rtu over tcp 和 modbus-tcp
[Description]: 1. 物联设备新增修改
2. modbus协议支持 modbus-rtu over tcp 和 modbus-tcp
[Root Cause]:
[Solution]:
[Test Suggestion]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/iot/entity/TIotDevice.java
@@ -104,7 +104,7 @@ public class TIotDevice implements Serializable {
     /**
      * 系数
      */
-    private Integer magnificationCoefficient;
+    private Float magnificationCoefficient;
 
     /**
      * 采集频率 单位:分钟
